check return value of libzfs_init() to make sure it works
libzfs_init() success depends on /sys being mounted to test for module loaded, or else ability to execute /sbin/modprobe zfs. libshare installs a constructor function that runs before main() that calls libzfs_init() and libzfs_fini(). If /sys is not already mounted and /sbin/modprobe is also not available, libzfs_init() fails. This can be a problem for a libzfs program running as an /init program in initramfs because the init program will not get a chance to mount /sys before libzfs_init() tries to read it. So even with the modules builtin, it is possible for libzfs_init() to fail if it can't read a mounted /sys/module/zfs. The error prints when the constructor function runs before main() because /sys it not yet mounted, but by the point where foobarz_init calls libzfs_init() /sys is mounted and normally returns success.